Basic stub creation
Quick and dirty implementation of /stubs/create, including redirection
to the created stub after it's been made. That should be reevaluated when we
actually know how people end up admining stubs -- you probably want to stay on
the creation form so you can put a bunch in in a row

The templates are currently completely unstyled.