Configure ideawheel to have CI tests run by travis-ci.com
Adds a basic .travis.yml to the project so that travis can run automated
tests against ideawheel. This allows us to have a nice, not-my-machine
option for running the ideawheel tests.

Also updates .gitignore to ignore emacs editor files in case someone has
forgotten to include that in their global git ignores.