Updates to library dependencies pursuant to the latest crypto news
Today was a bit of a red letter day for crypto vulnerabilities being
disclosed. The full list that I'm concerned about is available on the
openssl security advisory list here:
http://openssl.org/news/secadv_20150319.txt

I'm not aware of any live deployments of ideawheel that need to be restarted
or patched, however, since these exploits allow for fairly trivial denial
of service even on development machines, I figured an abundance of caution
was in order.